It reduce runtime cpu usage of fastime daemon. Instead of getting the time from OS everytime, it calculates the time from last retrieval time and add the refresh interval of the daemon.