@kevindiu 2019-05-21
New:
BenchmarkFastime-4   	2000000000	         1.01 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         1.01 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         0.99 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         0.99 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         1.00 ns/op	       0 B/op	       0 allocs/op

Original: 
BenchmarkFastime-4   	2000000000	         1.02 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         0.99 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         0.98 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         0.98 ns/op	       0 B/op	       0 allocs/op
BenchmarkFastime-4   	2000000000	         1.00 ns/op	       0 B/op	       0 allocs/op


Performance did not decreased a lot, but I think the runtime cpu usage will be decreased.