@kpango 2019-05-22
closed by https://github.com/kpango/fastime/pull/17